Bug 1356504 - Apply a default style for console.group string. r=bgrins
Because we now parse the arguments of console.group like in console.log,
we are using Reps to display the group name. This implies that the Reps
styles are used, which is not something we want for simple text group names.
We add specificity to the existing group rule so we can have the default style back.

MozReview-Commit-ID: A2zrF8ElZGZ

UltraBlame original commit: 0a15bd769093aea709311db221e11a62fb2983c0