Bug 1355233 - quote unmatched open paren in function tokens correctly; r=gl
When rewriting in the rule view, the rewriter quotes unmatched open
parens to avoid some edits affecting the rest of the style sheet.  This
was bug 1321970.  However, that change didn't correctly handle the case
of a function token, where the paren in question appears at the end, not
the start, of the token.

MozReview-Commit-ID: GjA40M2KsvX

UltraBlame original commit: 5323eee5a384afe38f3bd235a58e1713914ef169