Bug 1355585 - Streamline the format of "handlers.json", align the implementation, and reorganize tests. r=mak
This patch significantly improves the test coverage for both the JSON and RDF back-ends. There is a clearer separation between tests using predefined data and tests for the injection of default handlers. The predefined data includes more significant property combinations, and the JSON is now formatted. Helper functions are renamed for clarity.

Functions like "exists" that have different paths for MIME types and protocols are now tested with both, while behaviors that have a single path are now only tested with MIME types for efficiency.

The file format is redesigned to be more compact, and all the data is normalized when saving instead of when loading. Duplicates are now handled correctly when saving.

MozReview-Commit-ID: JI4I1M0N3lq

UltraBlame original commit: bcf5a28b1873a470c8626aeabae23ff69a191146