Bug 1358586: Require comma after <angle>, in -webkit-linear-gradient() CSS syntax. r=heycam
The logic around here is a bit tricky, because -moz-linear-gradient() shares
this codepath. Here's how things were going wrong:

* The -moz-linear-gradient syntax allows an angle (e.g. "30deg") to be followed
  by an optional box-position (e.g. "top right") before eventually requiring a
  comma and then some color stops.
* In contrast, the -webkit-linear-gradient syntax does NOT allow a box-position
  there (unless the box-position is by itself, without an angle).
* So, for -webkit-linear-gradient, we (correctly) skipped the box-position
  parsing code, if we've seen an angle already.
* BUT: in skipping that code, we *also* inadvertantly skipped the code that
  enforces that we eventually see a comma before we get to the color stops.
  So, we incorrectly accept some -webkit-linear-gradient() expressions that are
  missing a comma that really should be mandatory.

So, this patch adds a special case for -webkit-linear-gradient that
preemptively bails if we get an angle that's not followed by a comma. (With
that, we can also simplify the condition around the box position parsing, too.)

MozReview-Commit-ID: B7MQLxqe6D7

UltraBlame original commit: a9c09e578cf321de2429687ae5ad903190ce7b0c