Bug 1356237 - Ship a copy of transport module in marionette. r=ato
DevTools is about to move out of mozilla-central and be released as an add-on.
So that these protocol files are going to be missing from the tree.
To accommodate that, we are doing a copy of them next to marionette.

MozReview-Commit-ID: 9PyhuwyZyXI

UltraBlame original commit: 16a917f9afbcf8f75191f8189d5abccbe03b1c6b