Bug 1360389 - Have ChromiumCDMParent inform CDMProxy of shutdown. r=gerald
The ChromiumCDMParent is informed of the shutdown of its plugin, so we can
use that to inform the CDMProxy that its connection to the CDM has been
severed. This means we shutdown cleanly if the browser closes while playing.

MozReview-Commit-ID: HphQ2exu1gj

UltraBlame original commit: 4573433fcf30cd55ad93a81e7cf76b92f31109cb