Bug 1336763 - Wait for the browser to say it can go back before going back in browser_grouped_shistory_dead_navigate.js test. r=mystor
This test was accidentally taking advantage of the fact that closing a tab will result
in a nested event loop while waiting for the permitUnload messages to be sent back and
forth from the content process. This meant that the message that tells the parent that
the browser (which is having its history set) can now go back had a chance to be received
by the parent.

With the patches in bug 1336763, we're no longer spinning the event loop if the closing
tab doesn't have a beforeunload event handler in it, so we need to wait for the browser
to report that it can go back before actually sending it back in order to avoid a
test failure.

MozReview-Commit-ID: Lpl55iErrvf

UltraBlame original commit: c30cfe319b6d87ab1e7dacdde0320f23ada14536