Fix an APZ crash when a tab loads after its widget has been destroyed. (bug 1360478, r=kats)
Since tabs are loaded and shutdown asynchronously with respect to the UI
process, it is possible for a tab to be loading as its parent widget and
compositor are shutting down. We protect PAPZCTreeManager's constructor
against this by making a temporary APZCTM object.

UltraBlame original commit: e512947115142b8db267a902bc22e6c908d9e4a2