Bug 1360863 - Unremove emeNotifications.drmContentDisabled.message. r=dao
This string was (accidentally I presume) removed in bug 1330001, but it's
needed for the "enable DRM" prompt that we show if we visit Netflix or another
site that uses EME and EME is disabled.

Without this, users on Linux (where we ship EME disabled by default) will never
be prompted to turn on EME, and so Netflix will likely just fail with no
obvious way to make it work!

MozReview-Commit-ID: 9CoEQxb8Bir

UltraBlame original commit: 2ec98c2a48243d5671299f07a1d63396678a5c1f