Bug 1359879: add logging and robustness to key syncing, r=kmag
It seems like the server-side keyring is getting deleted somehow. This
causes syncing of the keyring to fail. We haven't completely
fixed the root problem yet, but this adds logging to try to flag
down the dangerous operation of deleting a bucket to make sure we
aren't doing that by accident. We also try to recover from the
server-keyring-was-deleted situation by wiping all storage.sync data,
the same as if we found a keyring that we couldn't decrypt.

MozReview-Commit-ID: JMB0IxApbGq

UltraBlame original commit: c06d46d320545f2ee43bc3b99ce19bea6e1e5d69