Bug 1314147 - Port the libvpx mozbuild generator to aom. r=froydnj,kinetik
This is a port of the libvpx scripts, themselves a port of
Chromium's scripts to generate an external build description
using hooks in the upstream configure and make scripts.
The libaom library is a fork of libvpx so we can use a similar
approach.

I've put the upstream source in $(topsrc_dir)/third_party/aom
but the build description and any patches are under the media
directory with the other codecs, similar to how zlib works.

Config files and headers generated by the upstream build system
are also under $(topsrc_dir)/media/libaom/.

MozReview-Commit-ID: ATLgOTPD2i1

UltraBlame original commit: 3398e932771022bd827367d5409d1457d66381c1