Bug 1360571 - Remove "local" bashism from mozconfig_loader. r=glandium
mozconfig_loader is invoked with /bin/sh, which may not be bash. We could
force mozconfigs to be run with bash instead, but that might be disruptive
for existing mozconfigs, and the sole bashism used in mozconfig_loader is
rather straightforward to workaround by namespacing the variables.

UltraBlame original commit: 6baa8102c70d55ad18c93728f270663ad6513d77