Bug 1354082 - part 1: make toggling the photon structure pref at runtime not break the panels, r=mikedeboer
This is unfortunate, but in order to keep automated tests working when we flip the pref,
it would be nice to be able to flip the pref at runtime and have things Just Work. This
tries to make that happen. We can remove most of this code post-Photon.

MozReview-Commit-ID: 8zbgCGJautO

UltraBlame original commit: 59cd845e4cb03cb40a88115017bfed937c1eb15e