Bug 1354082 - part 2: make overflow panel show up in customize mode and make dragging work correctly, r=mikedeboer
This reuses the logic we have for toolbar dragging and uses it for the photon panel (ie overflow panel)
to do vertical space-making animations.

It explicitly doesn't address the styling or permanently showing the overflow button (for which we have other bugs on file).


MozReview-Commit-ID: 4qrWC0H30xi

UltraBlame original commit: 119ce372e28e51d938596458ea7669c827a064d5