Bug 1362364: Lazily load the certificate database into the add-ons manager. r=rhelmer
The lazy loading is a little more complex because we want this to be a constant
in the scope so extensions can't trivially replace it. This also changes the
test to be more like the proof of concept from bug 1244248.

MozReview-Commit-ID: 2P890uRY1Si

UltraBlame original commit: 508966268fab932a33c9c77be039b6a1b8d3d4ec