Bug 1348919 - forced grid cell infobar and area infobar position; r=gl
The infobars needs to be not hidden when the `moveInfobar` function is called,
otherwise the computed style won't work as expected; therefore I switched the
functions' call related to that.
I added an optional `options` argument to the `moveInfobar` function in order
to support forced position and make the infobar hide if offscreen.

MozReview-Commit-ID: 81dxkMGt7vT

UltraBlame original commit: 161b368464fbbcdf1b9d8778a86709950eac8891