Bug 1355890 - Add Marionette XPCOM interface r=maja_zf
This adds a minimal XPCOM interface to access Marionette as a service.
All it does for now is to expose whether the Marionette server is
running, which will report true when the TCP listener socket is open and
false otherwise.

This will be used in browser/base/content/browser.js to determine
whether or not to add a visual UX cue on newly started <xul:browser>s.

MozReview-Commit-ID: 4Q9Oy2B9GQ1

UltraBlame original commit: 3d995af23408965b2f45a70859403ca001f8a752