Bug 1362364: Lazily load the certificate database into the add-ons manager. r=rhelmer
The lazy loading is a little more complex because we want this to be a constant
in the scope so extensions can't trivially replace it. This also changes the
test to be more like the proof of concept from bug 1244248.

I took the opportunity to promisify a bunch of things which ultimately made the
verifySignatures code nicer.

MozReview-Commit-ID: 2P890uRY1Si

UltraBlame original commit: 997a22336e37ce7892960a1ed6a0408be44b8b45