Bug 1351358 Part 1: Add flag to nsISHEntry to indicate if it was originally loaded in this process. r=bz
This flag is for when we've loaded a URI in a remote type that is not the default
for compatibility reasons (for example related http in the file content process).
So that we can load the history entry in that same process as well.

UltraBlame original commit: ff171bde1493a65250d8bb16ac7f7e60755e8738